% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transitionMatrix.R
\name{calculate_transitionMatrix}
\alias{calculate_transitionMatrix}
\title{Calculation of the transition matrix.}
\usage{
calculate_transitionMatrix(M, t, used_timepoints, tau)
}
\arguments{
\item{M}{Matrix containing all experimental data: the initial experimental setup matrix and the experimental cell state distribution matrices.}

\item{t}{This A vector containing all timepoints at which the cell state distributions have been experimentally measured. ---> realTime}

\item{used_timepoints}{A vector with the timepoints utilized to estimate the transition matrix. ---> Tau}
}
\description{
This function derives the transition matrix which contains the transition probabilitites between the distinct cell states.
}
\keyword{Matrix}
\keyword{matrix,}
\keyword{process}
\keyword{transition}
