% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltrans_plot.R
\name{celltrans_plot}
\alias{celltrans_plot}
\title{Allows to plot data and predictions.}
\usage{
celltrans_plot(input)
}
\arguments{
\item{input}{A datalist generated by the function readExperimentaldata.}
}
\description{
This function plots the predictions of CellTrans and the experimental data. The variable input contains the read data from readExperimentalData(). The user of CellTrans can choose in a dialog box the number of the experiment for which the plot shall be created.
}
\examples{
celltrans_plot
}
\keyword{plot}

