% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltrans_plotPDF.R
\name{celltrans_plotPDF}
\alias{celltrans_plotPDF}
\title{Plotting data points and the corresponding predictions of CellTrans in pdf format.}
\usage{
celltrans_plotPDF(input)
}
\arguments{
\item{input}{A datalist generated by the function readExperimentaldata().}
}
\description{
This function allows to create plots of the predictions of CellTrans together with the experimental data in pdf format. The variable input contains the read data from the function readExperimentalData(). One  can choose the experiments which are plotted as well as the directory and the filename.
}
\examples{
celltrans_plotPDF
}
\keyword{plot}

