% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltransitions.R
\name{celltransitions}
\alias{celltransitions}
\title{Main function of CellTrans: prints transition matrix and predicted equilibrium.}
\usage{
celltransitions(input)
}
\arguments{
\item{input}{A list with all necessary data for celltrans which can be generated by readExperimentalData().}
}
\description{
This functions derives and prints the transition probabilities and the predicted equilibrium distribution. The variable input contains the read data from the function readExperimentalData().
}
\keyword{Matrix}
\keyword{matrix,}
\keyword{process}
\keyword{transition}

