% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTrMatrix.R
\name{isTrMatrix}
\alias{isTrMatrix}
\title{Check for transition matrix.}
\usage{
isTrMatrix(A)
}
\arguments{
\item{A}{The matrix that is checked to be a transition matrix.}
}
\description{
Verification whether a matrix is a stochastic matrix of a Markov process or not, i.e. with non-negative entries and row sums equal to one.
}
\keyword{Matrix}
\keyword{matrix,}
\keyword{process}
\keyword{transition}

