% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QOM.R
\name{QOM}
\alias{QOM}
\title{Quasi-optimization of a matrix root to a stochastic matrix.}
\usage{
QOM(A)
}
\arguments{
\item{A}{The fractional root of a matrix that shall be regularized.}
}
\description{
QOM finds a transition matrix that is as close as possible, with respect to the Euclidean distance of the rows, to the fractional root of a given transition matrix. If the argument is a transition matrix, this matrix is returned unchanged.
}
\keyword{QOM,}
\keyword{matrix}
\keyword{regularization}
\keyword{root}

