% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_server.R
\name{shiny_server}
\alias{shiny_server}
\title{CellTrans2 Shiny App Server}
\usage{
shiny_server(input, output, session)
}
\description{
Server function for the CellTrans2 Shiny App, handling dynamic input fields,
data loading, transition matrix calculations, and result visualization.
}
