% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ui.R
\docType{data}
\name{shiny_ui}
\alias{shiny_ui}
\title{CellTrans2 Shiny App UI}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
shiny_ui
}
\description{
UI definition for the CellTrans2 Shiny App, including input widgets for configuring the analysis.
}
\keyword{datasets}
