% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transitionMatrix.R
\name{calculate_transitionMatrix}
\alias{calculate_transitionMatrix}
\title{Calculates a transition matrix.}
\usage{
calculate_transitionMatrix(M, t, used_timepoints)
}
\arguments{
\item{M}{Matrix containing all experimental data: the initial experimental setup matrix and the experimental cell state distribution matrices.}

\item{t}{This parameter is a vector with all timepoints at which the cell state distributions have been experimentally measured.}

\item{used_timepoints}{A vector with the timepoints that are used to estimate the transition matrix.}
}
\description{
This function derives the transition matrix which contains the transition probabilitites between the distinct cell states.
}
\keyword{Matrix}
\keyword{matrix,}
\keyword{process}
\keyword{transition}

