% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltrans_plot.R
\name{celltrans_plot}
\alias{celltrans_plot}
\title{Plotting data points and the corresponding predictions of CellTrans.}
\usage{
celltrans_plot(input)
}
\arguments{
\item{input}{A datalist generated by the function readExperimentaldata().}
}
\description{
This function allows to create plots of the predictions of CellTrans together with the experimental data. The variable input contains the read data from readExperimentalData(). One can choose the experiments for which the plot is created.
}
\examples{
celltrans_plot
}
\keyword{plot}

