% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTrMatrix.R
\name{isTrMatrix}
\alias{isTrMatrix}
\title{Is a matrix a transition matrix?}
\usage{
isTrMatrix(A)
}
\arguments{
\item{A}{The matrix that is checked to be a transition matrix.}
}
\description{
A function that checks whether a matrix is a stochastic matrix describing a Markov process, i.e. with non-negative entries and row sums equal to one.
}
\keyword{Matrix}
\keyword{matrix,}
\keyword{process}
\keyword{transition}

