% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExperimentalData.R
\name{readExperimentalData}
\alias{readExperimentalData}
\title{Reads all data for CellTrans.}
\usage{
readExperimentalData()
}
\description{
This function reads all necessary data. First, it opens a dialog box which asks for the number of cell types, the names of the cell types, the time-step length and the time points of measurement. Then, the files containing the cell state proportion matrices are read. First, the initial experimental setup matrix can either be chosen as identity matrix (for pure initial cell populations) or a custom initial matrix can be provided. Then, the experimental cell proportion matrices are read for each time point of measurement. It is recommendable to save the input into a variable for the further analysis, e.g. input <- readExperimentalData().
}
\examples{
input<-readExperimentalData()
}
\keyword{cell}
\keyword{distribution}
\keyword{experimental}
\keyword{initial}
\keyword{matrices}
\keyword{matrix,}

